go <1.7 doesn't support all current fields
